# perf(npm-resolver): improve performance of semver.satisfies