--- a/.changeset/pretty-toys-agree.md
+++ b/.changeset/pretty-toys-agree.md
@@ -1,0 +1,5 @@
+---
+"@pnpm/npm-resolver": patch
+---
+
+Improve performance of clean install by preconverting and caching semver objects