@jakebailey 2023-03-31
It is probably better to yank this out into a shared module somewhere, however, it doesn't seem like this code is hot anywhere else and there didn't seem to be a good place to put it. I wasn't about to go and add a new package to the repo.