@zkochan 2023-03-31
oh wow, this is awesome!

> Even faster was to cache the comparison itself

How much faster?