@jakebailey 2023-03-31
Let me recreate that change and I'll let you know. it wasn't that much better, but it was a somewhat peak memory usage increase.