@jakebailey 2023-03-31
Here's main now that this PR is merged:

```
Done in 2m 13.1s
total time:  133.38s
user time:   161.73s
system time: 31.03s
CPU percent: 144%
max memory:  1669 MB
```

And, if I just put a cache on top:

```
Done in 1m 49.2s
total time:  109.50s
user time:   122.74s
system time: 31.50s
CPU percent: 140%
max memory:  1989 MB
```

If https://github.com/npm/node-semver/pull/528 is merged upstream, then the final performance will be about the same as the full caching but without the 300MB+ peak memory increase, so, I'm hoping for that.

For posterity, here is the function I tested:

```ts
const semverSatisfiesCache = new Map<string, boolean>()

// This is a performance optimization; working with string-ish semver
// causes lots of allocations and repeated work, but caching the Range
// and ensuring we give it a SemVer instance greatly speeds things up.
function semverSatisfiesLoose (version: string, range: string): boolean {
  const key = `${version},${range}`
  let result = semverSatisfiesCache.get(key)
  if (result !== undefined) {
    return result
  }

  try {
    const semverRange = new semver.Range(range, true)
    const semverVersion = new semver.SemVer(version, true)
    result = semverRange.test(semverVersion)
  } catch {
    result = false
  }

  semverSatisfiesCache.set(key, result)
  return result
}
```