@jakebailey 2023-03-31
To be clear, the fork/PR is frin @H4ad; I didn't do anything but point out that `parseOptions` function!