@weyert 2023-03-31
Amazing work @jakebailey. I appreciate your hard work to make PNPM faster. I am excited when TD switches to it. My former colleague John Reilly introduced me to TD many years ago :)